+++ backend/cmd/api/dtos.go
@@ -1,8 +1,9 @@
 package main
 
 type FileItem struct {
-	Name  string `json:"name"`
-	IsDir bool   `json:"isDirectory"`
+	Name        string `json:"name"`
+	IsDir       bool   `json:"isDirectory"`
+	IsSupported bool   `json:"isSupported"`
 }
 
 type FileList struct {
@@ -14,4 +15,8 @@ type AudioMetadata struct {
 	Codec      string  `json:"codec"`
 	SampleRate int     `json:"sampleRate"` // in Hz
 	Duration   float64 `json:"duration"`   // in seconds
+}
+
+type Features struct {
+	EnableLogScale bool `json:"enableLogScale"`
 }