+++ backend/cmd/api/routes.go
@@ -1,31 +1,22 @@
 package main
 
 import (
-	"embed"
-	"io/fs"
 	"net/http"
 
 	"github.com/julienschmidt/httprouter"
 )
 
-//go:embed ui/*
-var uiFiles embed.FS
-
-func (app *application) routes() (http.Handler, error) {
-	uiFS, err := fs.Sub(uiFiles, "ui")
-	if err != nil {
-		return nil, err
-	}
-
+func (app *application) routes() http.Handler {
 	router := httprouter.New()
 
-	router.NotFound = http.FileServerFS(uiFS)
+	router.NotFound = http.FileServer(http.Dir("./web"))
 
 	router.HandlerFunc(http.MethodGet, "/v1/files", app.listFilesHandler)
 	router.HandlerFunc(http.MethodGet, "/v1/audio", app.getAudioFileHandler)
 	router.HandlerFunc(http.MethodGet, "/v1/audio/metadata", app.getAudioMetadataHandler)
+	router.HandlerFunc(http.MethodGet, "/v1/features", app.getFeaturesHandler)
 
 	router.MethodNotAllowed = http.HandlerFunc(app.methodNotAllowedResponse)
 
-	return app.recoverPanic(router), nil
+	return app.recoverPanic(router)
 }