+++ frontend/src/core/FileItem.ts
@@ -1,4 +1,5 @@
 export default interface FileItem {
   name: string;
   isDirectory: boolean;
+  isSupported: boolean;
 }